Twiml

Documentation

class Twiml p t | t -> p

TwiML is a set of instructions you can use to tell Twilio what to do when you receive an incoming call or SMS. See https://www.twilio.com/docs/api/twiml.

This library provides a number of smart constructors for creating Twiml, as well as conversion functions for interop with other Haskell XML libraries, including HXT.

As an example, the following Haskell code,

example
  = respond
  . sayMan "Hello, world"
  $ hangup

Is transformed into,

<?xml version="1.0" encoding="UTF-8"?>
<Response>
  <Say voice="man">Hello, world</Say>
  <Hangup/>
</Response>

Types

<Response>

data Response

The root element of Twilio's XML Markup is the <Response> element. See https://www.twilio.com/docs/api/twiml/your_response#response-element.

respond :: Twiml Response t => t -> Response

toArrowXml :: ArrowXml a => Response -> a n XmlTree

URL

data URL

Method

data Method

Constructors

GET

POST

Key

data Key

Constructors

K0
0

K1
1

K2
2

K3
3

K4
4

K5
5

K6
6

K7
7

K8
8

K9
9

KStar
*

KPound
#

End

data End p

End is not a TwiML verb, but it is useful for terminating Twiml.

end :: End p

Primary Verbs

<Say>

data Say p

The <Say> verb converts text to speech that is read back to the caller. See https://www.twilio.com/docs/api/twiml/say.

data SayAttributes

See https://www.twilio.com/docs/api/twiml/say#attributes.

Constructors

SayAttributes

Fields

sayVoice :: Maybe Voice

sayLoop :: Maybe Natural

data Voice

Voices supported by <Say>. See https://www.twilio.com/docs/api/twiml/say#attributes-voice.

Constructors

Man (Maybe Lang)

Woman (Maybe Lang)

Alice (Maybe LangAlice)

data Lang

Languages spoken by voices Man and Woman. See https://www.twilio.com/docs/api/twiml/say#attributes-manwoman.

Constructors

English

EnglishUK

Spanish

French

German

Italian

data LangAlice

Languages spoken by Alice. See https://www.twilio.com/docs/api/twiml/say#attributes-alice.

Constructors

DaDK
Danish, Denmark

DeDE
German, Germany

EnAU
English, Australia

EnCA
English, Canada

EnGB
English, UK

EnIN
English, India

EnUS
English, United States

CaES
Catalan, Spain

EsES
Spanish, Spain

EsMX
Spanish, Mexico

FiFI
Finnish, Finland

FrCA
French, Canada

FrFR
French, France

ItIT
Italian, Italy

JaJP
Japanese, Japan

KoKR
Korean, Korea

NbNO
Norwegian, Norway

NlNL
Dutch, Netherlands

PlPL
Polish-Poland

PtBR
Portuguese, Brazil

PtPT
Portuguese, Portugal

RuRU
Russian, Russia

SvSE
Swedish, Sweden

ZhCN
Chinese (Mandarin)

ZhHK
Chinese (Cantonese)

ZhTW
Chinese (Taiwanese Mandarin)

say :: Twiml p t => String -> t -> Say p

say' :: Twiml p t => SayAttributes -> String -> t -> Say p

sayMan :: Twiml p t => String -> t -> Say p

sayMan' :: Twiml p t => Lang -> String -> t -> Say p

sayWoman :: Twiml p t => String -> t -> Say p

sayWoman' :: Twiml p t => Lang -> String -> t -> Say p

sayAlice :: Twiml p t => String -> t -> Say p

sayAlice' :: Twiml p t => LangAlice -> String -> t -> Say p

voice :: Lens (Say p) (Say p) (Maybe Voice) Voice

sayAttributes :: Lens' (Say p) SayAttributes

<Play>

data Play p

The <Play> verb plays an audio file back to the caller. Twilio retrieves the file from a URL that you provide. See https://www.twilio.com/docs/api/twiml/play.

data PlayAttributes

See https://www.twilio.com/docs/api/twiml/play#attributes.

Constructors

PlayAttributes

Fields

playLoop :: Maybe Natural

play :: Twiml p t => URL -> t -> Play p

play' :: Twiml p t => PlayAttributes -> URL -> t -> Play p

playAttributes :: Lens' (Play p) PlayAttributes

<Gather>

data Gather p

The <Gather> verb collects digits that a caller enters into his or her telephone keypad. See https://www.twilio.com/docs/api/twiml/gather.

You can nest <Say>, <Play>, and <Pause> in <Gather>. See https://www.twilio.com/docs/api/twiml/gather#nesting-rules.

Nesting rules are made explicit via an unexported functor, GatherNoun.

data GatherAttributes

See https://www.twilio.com/docs/api/twiml/gather#attributes.

Constructors

GatherAttributes

Fields

gatherAction :: Maybe URL

gatherMethod :: Maybe Method

gatherTimeout :: Maybe Natural

gatherFinishOnKey :: Maybe Key

gatherNumDigits :: Maybe Natural

gather :: (Twiml GatherNoun n, Twiml p t, NotGatherNoun p) => n -> t -> Gather p

gather' :: (Twiml GatherNoun n, Twiml p t, NotGatherNoun p) => GatherAttributes -> n -> t -> Gather p

numDigits :: Lens (Gather p) (Gather p) (Maybe Natural) Natural

gatherAttributes :: Lens' (Gather p) GatherAttributes

<Record>

data Record p

The <Record> verb records the caller's voice and returns to you the URL of a file containing the audio recording. See https://www.twilio.com/docs/api/twiml/record.

data RecordAttributes

See https://www.twilio.com/docs/api/twiml/record#attributes.

Constructors

RecordAttributes

Fields

recordAction :: Maybe URL

recordMethod :: Maybe Method

recordTimeout :: Maybe Natural

recordFinishOnKey :: Maybe Key

recordMaxLength :: Maybe Natural

recordTranscribe :: Maybe Bool

recordTranscribeCallback :: Maybe URL

recordPlayBeep :: Maybe Bool

record :: (Twiml p t, NotGatherNoun p) => URL -> t -> Record p

record' :: (Twiml p t, NotGatherNoun p) => RecordAttributes -> URL -> t -> Record p

maxLength :: Lens (Record p) (Record p) (Maybe Natural) Natural

transcribe :: Lens (Record p) (Record p) (Maybe Bool) Bool

transcribeCallback :: Lens (Record p) (Record p) (Maybe URL) URL

playBeep :: Lens (Record p) (Record p) (Maybe Bool) Bool

recordAttributes :: Lens' (Record p) RecordAttributes

<Sms>

data Sms p

The <Sms> verb sends an SMS message to a phone number during a phone call. See https://www.twilio.com/docs/api/twiml/sms.

data SmsAttributes

See https://www.twilio.com/docs/api/twiml/sms#attributes.

Constructors

SmsAttributes

Fields

smsTo :: Maybe String

smsFrom :: Maybe String

smsAction :: Maybe URL

smsMethod :: Maybe Method

smsStatusCallback :: Maybe URL

sms :: (Twiml p t, NotGatherNoun p) => String -> t -> Sms p

sms' :: (Twiml p t, NotGatherNoun p) => SmsAttributes -> String -> t -> Sms p

to :: Lens (Sms p) (Sms p) (Maybe String) String

from :: Lens (Sms p) (Sms p) (Maybe String) String

statusCallback :: Lens (Sms p) (Sms p) (Maybe URL) URL

smsAttributes :: Lens' (Sms p) SmsAttributes

<Dial>

data Dial p

The <Dial> verb connects the current caller to another phone. See https://www.twilio.com/docs/api/twiml/dial.

data DialAttributes

See https://www.twilio.com/docs/api/twiml/dial#attributes.

Constructors

DialAttributes

Fields

dialAction :: Maybe URL

dialMethod :: Maybe Method

dialTimeout :: Maybe Natural

dialHangupOnStar :: Maybe Bool

dialTimeLimit :: Maybe Natural

dialCallerId :: Maybe String

dialRecord :: Maybe Bool

data DialNoun

dial :: (Twiml p t, NotGatherNoun p) => Either DialNoun String -> t -> Dial p

dial' :: (Twiml p t, NotGatherNoun p) => DialAttributes -> Either DialNoun String -> t -> Dial p

hangupOnStar :: Lens (Dial p) (Dial p) (Maybe Bool) Bool

timeLimit :: Lens (Dial p) (Dial p) (Maybe Natural) Natural

callerId :: Lens (Dial p) (Dial p) (Maybe String) String

recordA :: Lens (Dial p) (Dial p) (Maybe Bool) Bool

dialAttributes :: Lens' (Dial p) DialAttributes

Secondary Verbs

<Enqueue>

data Enqueue p

The <Enqueue> verb enqueues the current call in a call queue. See https://www.twilio.com/docs/api/twiml/enqueue.

enqueue :: (Twiml p t, NotGatherNoun p) => String -> t -> Enqueue p

<Leave>

data Leave p

The <Leave> verb transfers control of a call that is in a queue so that the caller exits the queue and execution continues with the next verb after the original <Enqueue>. See https://www.twilio.com/docs/api/twiml/leave.

leave :: NotGatherNoun p => Leave p

<Hangup>

data Hangup p

The <Hangup> verb ends a call. See https://www.twilio.com/docs/api/twiml/hangup.

hangup :: NotGatherNoun p => Hangup p

<Redirect>

data Redirect p

The <Redirect> verb transfers control of a call to the TwiML at a different URL. See https://www.twilio.com/docs/api/twiml/redirect.

data RedirectAttributes

See https://www.twilio.com/docs/api/twiml/redirect#attributes.

Constructors

RedirectAttributes

Fields

redirectMethod :: Maybe Method

redirect :: NotGatherNoun p => URL -> Redirect p

redirect' :: NotGatherNoun p => RedirectAttributes -> URL -> Redirect p

redirectAttributes :: Lens' (Redirect p) RedirectAttributes

<Reject>

data Reject p

The <Reject> verb rejects an incoming call to your Twilio number without billing you. See https://www.twilio.com/docs/api/2010-04-01/twiml/reject.

data RejectAttributes

See https://www.twilio.com/docs/api/twiml/reject#attributes.

Constructors

RejectAttributes

Fields

rejectReason :: Maybe Reason

data Reason

The reason attribute takes the values "rejected" and "busy." This tells Twilio what message to play when rejecting a call. Selecting "busy" will play a busy signal to the caller, while selecting "rejected" will play a standard not-in-service response. See https://www.twilio.com/docs/api/twiml/reject#attributes-reason

reject :: NotGatherNoun p => Reject p

reject' :: NotGatherNoun p => RejectAttributes -> Reject p

reason :: Lens (Reject p) (Reject p) (Maybe Reason) Reason

rejectAttributes :: Lens' (Reject p) RejectAttributes

<Pause>

data Pause p

The <Pause> verb waits silently for a specific number of seconds. See https://www.twilio.com/docs/api/twiml/pause.

data PauseAttributes

See https://www.twilio.com/docs/api/twiml/pause#attributes.

Constructors

PauseAttributes

Fields

pauseLength :: Maybe Natural

pause :: Twiml p t => t -> Pause p

pause' :: Twiml p t => PauseAttributes -> t -> Pause p

length :: Lens (Pause p) (Pause p) (Maybe Natural) Natural

pauseAttributes :: Lens' (Pause p) PauseAttributes

Lenses

class HasLoop t

loop :: HasLoop t => Lens t t (Maybe Natural) Natural

class HasAction t

action :: HasAction t => Lens t t (Maybe URL) URL

class HasMethod t

method :: HasMethod t => Lens t t (Maybe Method) Method

class HasTimeout t

timeout :: HasTimeout t => Lens t t (Maybe Natural) Natural

class HasFinishOnKey t

finishOnKey :: HasFinishOnKey t => Lens t t (Maybe Key) Key