Twiml

Documentation

class Twiml t

TwiML is a set of instructions you can use to tell Twilio what to do when you receive an incoming call or SMS. See https://www.twilio.com/docs/api/twiml.

This library provides a number of smart constructors for creating Twiml, as well as conversion functions for interop with other Haskell XML libraries, including HXT.

As an example, the following Haskell code,

example
  = response
  . sayMan "Hello, world"
  . sayWoman "Goodbye, world"
  $ hangup

Is transformed into,

<?xml version="1.0" encoding="UTF-8"?>
<Response>
  <Say voice="man">Hello, world</Say>
  <Say voice="woman">Goodbye, world</Say>
  <Hangup/>
</Response>

data Null

Useful for terminating Twiml.

data Response

The root element of Twilio's XML Markup is the <Response> element. In any TwiML response to a Twilio request, all verb elements must be nested within this element. Any other structure is considered invalid. See https://www.twilio.com/docs/api/twiml/your_response#response-element.

respond :: Twiml t => (forall a. f a -> a) -> f t -> Response

toArrowXml :: ArrowXml a => Response -> a n XmlTree

Convert a Response into an ArrowXml to be used with HXT. See http://hackage.haskell.org/package/hxt.

Primary Verbs

<Say>

data Say

The <Say> verb converts text to speech that is read back to the caller See https://www.twilio.com/docs/api/twiml/say.

data Voice

Voices supported by <Say>. See https://www.twilio.com/docs/api/twiml/say#attributes-voice.

Constructors

Man (Maybe Lang)

Woman (Maybe Lang)

Alice (Maybe LangAlice)

data Lang

Languages spoken by voices Man and Woman. See https://www.twilio.com/docs/api/twiml/say#attributes-manwoman.

Constructors

English

EnglishUK

Spanish

French

German

Italian

data LangAlice

Languages spoken by Alice. See https://www.twilio.com/docs/api/twiml/say#attributes-alice.

Constructors

DaDK
Danish, Denmark

DeDE
German, Germany

EnAU
English, Australia

EnCA
English, Canada

EnGB
English, UK

EnIN
English, India

EnUS
English, United States

CaES
Catalan, Spain

EsES
Spanish, Spain

EsMX
Spanish, Mexico

FiFI
Finnish, Finland

FrCA
French, Canada

FrFR
French, France

ItIT
Italian, Italy

JaJP
Japanese, Japan

KoKR
Korean, Korea

NbNO
Norwegian, Norway

NlNL
Dutch, Netherlands

PlPL
Polish-Poland

PtBR
Portuguese, Brazil

PtPT
Portuguese, Portugal

RuRU
Russian, Russia

SvSE
Swedish, Sweden

ZhCN
Chinese (Mandarin)

ZhHK
Chinese (Cantonese)

ZhTW
Chinese (Taiwanese Mandarin)

sayMan :: (Functor f, Twiml t) => String -> f t -> f Say

sayMan' :: (Functor f, Twiml t) => Lang -> String -> f t -> f Say

sayWoman :: (Functor f, Twiml t) => String -> f t -> f Say

sayWoman' :: (Functor f, Twiml t) => Lang -> String -> f t -> f Say

sayAlice :: (Functor f, Twiml t) => String -> f t -> f Say

sayAlice' :: (Functor f, Twiml t) => LangAlice -> String -> f t -> f Say

<Play>

data Play

The <Play> verb plays an audio file back to the caller. Twilio retrieves the file from a URL that you provide. See https://www.twilio.com/docs/api/twiml/play.

play :: (Functor f, Twiml t) => URL -> f t -> f Play

play' :: (Functor f, Twiml t) => Natural -> URL -> f t -> f Play

<Gather>

data Gather

The <Gather> verb collects digits that a caller enters into his or her telephone keypad. See https://www.twilio.com/docs/api/twiml/gather.

You can nest <Say>, <Play>, and <Pause> in <Gather>. See https://www.twilio.com/docs/api/twiml/gather#nesting-rules.

Nesting rules are made explicit via an unexported functor, GatherNoun.

gatherNull :: GatherNoun Null

gather :: (Functor f, f a :/~ GatherNoun a, Twiml n, Twiml t) => GatherNoun n -> f t -> f Gather

<Record>

data Record

The <Record> verb records the caller's voice and returns to you the URL of a file containing the audio recording. See https://www.twilio.com/docs/api/twiml/record.

record :: (Functor f, f a :/~ GatherNoun a, Twiml t) => f t -> f Record

<Sms>

data Sms

The <Sms> verb sends an SMS message to a phone number during a phone call. See https://www.twilio.com/docs/api/twiml/sms.

sms :: (Functor f, f a :/~ GatherNoun a, Twiml t) => String -> f t -> f Sms

<Dial>

data Dial

The <Dial> verb connects the current caller to another phone. See https://www.twilio.com/docs/api/twiml/dial.

data DialNoun

Secondary Verbs

<Enqueue>

data Enqueue

The <Enqueue> verb enqueues the current call in a call queue. See https://www.twilio.com/docs/api/twiml/enqueue.

<Leave>

data Leave

The <Leave> verb transfers control of a call that is in a queue so that the caller exits the queue and execution continues with the next verb after the original <Enqueue>. See https://www.twilio.com/docs/api/twiml/leave.

leave :: f a :/~ GatherNoun a => (forall a. a -> f a) -> f Leave

<Hangup>

data Hangup

The <Hangup> verb ends a call. See https://www.twilio.com/docs/api/twiml/hangup.

hangup :: f a :/~ GatherNoun a => (forall a. a -> f a) -> f Hangup

<Redirect>

data Redirect

The <Redirect> verb transfers control of a call to the TwiML at a different URL. See https://www.twilio.com/docs/api/twiml/redirect.

<Reject>

data Reject

The <Reject> verb rejects an incoming call to your Twilio number without billing you. See https://www.twilio.com/docs/api/2010-04-01/twiml/reject.

data Reason

The reason attribute takes the values rejected and busy. This tells Twilio what message to play when rejecting a call. Selecting busy will play a busy signal to the caller, while selecting rejected will play a standard not-in-service response. See https://www.twilio.com/docs/api/twiml/reject#attributes-reason

reject :: f a :/~ GatherNoun a => (forall a. a -> f a) -> f Reject

reject' :: f a :/~ GatherNoun a => (forall a. a -> f a) -> Reason -> f Reject

<Pause>

data Pause

The <Pause> verb waits silently for a specific number of seconds. See https://www.twilio.com/docs/api/twiml/pause.

pause :: (Functor f, f a :/~ GatherNoun a, Twiml t) => f t -> f Pause

pause' :: (Functor f, f a :/~ GatherNoun a, Twiml t) => Natural -> f t -> f Pause